Fade working again.  Overlaid tracked views working with async rendering.  Drawable atlases now respecting draw priority and visibility range.  Fixed a but in single byte compressed textures in texture atlases.  A few more changes related to the great double migration.  Now rescaling texture to include a boundary, thus fixing the boundary problem in texture atlases.  Dealing with view resize failing in EAGLView.  Loft layer now has a special mode for filled, layered polys.  Fixed an ordering problem with skirts.  Labels working again.

+++ WhirlyGlobeSrc/WhirlyGlobeLib/src/Drawable.mm
@@ -302,7 +302,7 @@ bool BasicDrawable::isOn(WhirlyKitRendererFrameInfo *frameInfo) const
     if (minVisible == DrawVisibleInvalid || !on)
         return on;
 
-    float visVal = [frameInfo.theView heightAboveSurface];
+    double visVal = [frameInfo.theView heightAboveSurface];
     
     return ((minVisible <= visVal && visVal <= maxVisible) ||
              (maxVisible <= visVal && visVal <= minVisible));
